fix :bug: fixed the int overflow for larger images

+++ face_detection/retinaface/prior_box.py
@@ -8,8 +8,8 @@ from math import ceil
 def generate_prior_box(feature_maps, image_size, steps, min_sizes):
     n_anchors = 0
     for x in feature_maps:
-        n_anchors += x[0] * x[1] * len(min_sizes[0])
-    anchors = np.empty((n_anchors*4), dtype=np.float32)
+        n_anchors += int(x[0]) * int(x[1]) * len(min_sizes[0])
+    anchors = np.empty((n_anchors*4), dtype=np.float64)
 #    print(feature_maps, image_size, steps, min_sizes)
     idx_anchor = 0
     for k, f in enumerate(feature_maps):
